requirements.txt
openai==0.28.0
pdfminer-six
python-docx
streamlit==1.30.0
Jinja2
docx2txt
streamlit-ext==0.1.9
stqdm==0.0.5
setuptools==68.2.2
# google-api-python-client google-auth-httplib2 google-auth-oauthlib
psycopg2-binary
sqlalchemy
